requirements.txt
matplotlib>=3.5
numpy>=1.20
pandas>=1.3
Pillow>=9.0
seaborn>=0.11

# Optional: for extended colormaps in colorbar
# cmasher>=1.6
